# src/static/swagger.yaml
openapi: 3.0.3
info:
  title: TrustManager API
  description: API for managing and distributing trusted certificates and truststores.
  version: "1.0.0"
servers:
  - url: http://localhost:5100

paths:
  /Certificate:
    put:
      summary: Add a new certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              description: PEM-encoded certificate
      responses:
        '201':
          description: Certificate added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input or parsing error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Certificates:
    get:
      summary: Get all certificates
      responses:
        '200':
          description: List of certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CertificateSummary'

  /Trust:
    post:
      summary: Trust a certificate (only self-signed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
      responses:
        '200':
          description: Certificate trusted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing certificate ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: Certificate is not self-signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Certificate not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Distrust:
    post:
      summary: Distrust a certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
      responses:
        '200':
          description: Certificate distrusted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing certificate ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Certificate not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Certificate/serial/{serial}:
    get:
      summary: Get certificate by serial number
      parameters:
        - name: serial
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PEM-encoded certificate
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Certificate not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Certificate/subject/{subject}:
    get:
      summary: Get certificate(s) by subject (full or partial)
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of PEM-encoded certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: No certificates found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Certificate/fingerprint/{fingerprint}:
    get:
      summary: Get certificate by fingerprint
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PEM-encoded certificate
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Certificate not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /Certificates/atom:
    get:
      summary: Get ATOM feed of trusted certificates
      responses:
        '200':
          description: ATOM XML feed
          content:
            application/atom+xml:
              schema:
                type: string

  /Truststore/{format}:
    get:
      summary: Download truststore file in specified format
      parameters:
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [jks, pfx, pem, rpm]
      responses:
        '200':
          description: Truststore file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Truststore file not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /BatchJob:
    post:
      summary: Run a batch job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job:
                  type: string
                  enum: [assemble_jks, assemble_pfx, assemble_pem, assemble_rpm]
                  description: Name of the batch job to run
      responses:
        '200':
          description: Batch job executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stdout:
                    type: string
                  stderr:
                    type: string
        '400':
          description: Invalid or missing job name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  available_jobs:
                    type: array
                    items:
                      type: string
        '500':
          description: Batch job failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  stdout:
                    type: string
                  stderr:
                    type: string

  /BatchJob/list:
    get:
      summary: List available batch jobs
      description: Returns a list of available batch job names that can be executed.
      responses:
        '200':
          description: List of batch job names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /Governance/Truststore:
    post:
      summary: Add a new governed truststore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                truststore_type:
                  type: string
                  enum: [JKS, PKCS12, CAPI, PEM File(s)]
                host:
                  type: string
                location:
                  type: string
                certificate_ids:
                  type: array
                  items:
                    type: integer
                notes:
                  type: string
      responses:
        '201':
          description: Truststore added successfully
        '400':
          description: Missing or invalid fields
        '500':
          description: Internal server error

  /Governance/Truststore/{truststore_id}/notes:
    post:
      summary: Append notes to a governed truststore
      parameters:
        - name: truststore_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        '200':
          description: Notes appended successfully
        '400':
          description: Missing or invalid fields
        '404':
          description: Truststore not found
        '500':
          description: Internal server error

components:
  schemas:
    CertificateSummary:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        serial:
          type: string
        fingerprint:
          type: string